requirements.txt
fastapi
uvicorn[standard]
jinja2
requests
pydantic
websocket-client
python-slugify
psycopg2-binary
